+++ apps/react-demo-app/testcafe/desktop/login.role.ts
@@ -7,9 +7,7 @@ export const regularUser = Role(
   process.env.BASE_URL,
   async (t) => {
     log('Attempting to perform fake IDP login');
-    const loginButton = screen.getByText(
-      'Login with Fake IDP (react-demo-app-e2e-azure-dev)'
-    );
+    const loginButton = screen.getByText('Login with Fake IDP (azure-dev)');
     await t.click(loginButton);
   },
   { preserveUrl: true }